requirements.txt
fastapi
uvicorn
primp
duckduckgo_search
python-dotenv
termcolor
janus
sse_starlette
class_registry
griffe
setuptools
sphinx
bs4
cachetools
streamlit
pyvis